Move collision flags into its own file
