+++ Assets/Sandbox/Minimal/Physics/CollisionFlags2D.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d21c02a012f6a0c4b98233ba16d5d4f4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
